requirements/base.txt
Django==5.1.2
django-cors-headers==4.5.0
django-decouple==2.1
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-shortcuts==1.6
djangorestframework==3.15.2
ipython==8.28.0
virtualenv==20.27.0
